[rename] search store 이름 수정
