+++ src/mocks/handlers.ts
@@ -1,6 +1,13 @@
+import kakaoHandler from './kakaoHandlers';
 import myPageHandlers from './myPageHandlers';
 import pointHandlers from './pointHandlers';
 import recruitHandlers from './recruitHandlers';
 import studyMainHandlers from './studyMainHandlers';
 
-export const handlers = [...recruitHandlers, ...myPageHandlers, ...pointHandlers, ...studyMainHandlers];
+export const handlers = [
+  ...recruitHandlers,
+  ...myPageHandlers,
+  ...pointHandlers,
+  ...studyMainHandlers,
+  ...kakaoHandler,
+];
